refactor nips relic code. WIP.
