+++ relic/vqa/collision_episodes_generation.py
