+++ relic/vqa/sanity_checker.py
