+++ relic/vqa/scene_level_functionals.py
