+++ relic/vqa/scratch.py
