+++ relic/vqa/static_question_generator.py
@@ -1,10 +1,10 @@
-from typing import Any, Iterable, Union, LiteralString, Callable
+from typing import Any, Iterable, Union, Callable
 from vqa.functionals import color_wrapper, type_wrapper, state_wrapper, action_wrapper, pos_wrapper, count, \
     CountGreater, CountEqual, Identity, locate_wrapper, extract_color, extract_type, extract_color_unique, \
     extract_type_unique
 from vqa.scene_graph import SceneGraph
 from vqa.object_node import ObjectNode
-from vqa.grammar import STATIC_GRAMMAR
+from relic.vqa.grammar import STATIC_GRAMMAR
 import random
 import json
 import argparse
